bcachefs: new "debug" and "list_bkeys" commands
This introduces two new commands to the bcachefs utility. The first,
"debug", is used for directly manipulating bkeys in the underlying
btrees. It has two subcommands, "dump" which takes a btree and bpos and
prints the data from that bkey, and "update" which is used to update a
given field of a key to a value. For example:

$ bcachefs debug ~/test-img -c "dump inodes 0:4096:U32_MAX"
u64s 17 type inode_v3 0:4096:U32_MAX len 0 ver 0:   mode=40755
  flags= (16300000)
  journal_seq=9
  ...

$ bcachefs debug ~/test-img -c "update inodes 0:4096:U32_MAX bch_inode_unpacked.bi_nlink=2"

The second command, "list_bkeys", is used to list the known bkey types
as well as their fields. The types and fields listed are the ones that
can be accessed by the "debug" command.

Signed-off-by: Thomas Bertschinger <tahbertschinger@gmail.com>