do not set "read_only" mount option before fsck
The filesystem option "read_only" was added as a mount option, but later
removed because it is redundant with the standard "ro" option.

Rather than passing "read_only" as a mount option in the fsck ioctl
prior to initiating a kernel-side fsck, bcachefs now simply sets that
option in the kernel before beginning offline fsck.

Signed-off-by: Thomas Bertschinger <tahbertschinger@gmail.com>